+++ src/routes/login.jsx
@@ -6,7 +6,7 @@
 'use client'
 
 import {useState} from "react";
-import AuthService from "@/services/login/index.js";
+import AuthService from "@/services/login/login.js";
 import {useNavigate} from "react-router-dom";
 import {useForm} from "react-hook-form";
 import {LoginForm} from "@/components/loginForm.jsx";
